# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borehole Drilling Services Kenya API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  slug: string;
  category: string;
  shortDescription: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  /** @nullable */
  priceFrom?: number | null;
  /** @nullable */
  priceUnit?: string | null;
  /** @nullable */
  duration?: string | null;
  highlights?: string[];
  coverageAreas?: string[];
}

export type ProductSpecifications = {[key: string]: string};

export interface Product {
  id: number;
  name: string;
  slug: string;
  category: string;
  /** @nullable */
  brand?: string | null;
  description: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  inStock: boolean;
  /** @nullable */
  stockQty?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  specifications?: ProductSpecifications;
  deliveryAvailable?: boolean;
  /** @nullable */
  deliveryDays?: number | null;
  whatsappOrderEnabled?: boolean;
}

export interface ProductPage {
  items: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface CategorySummary {
  name: string;
  count: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImageUrl?: string | null;
  price: number;
  quantity: number;
  subtotal: number;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
  itemCount?: number;
}

export interface CartItemInput {
  sessionId: string;
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export type InquiryInputInquiryType = typeof InquiryInputInquiryType[keyof typeof InquiryInputInquiryType];


export const InquiryInputInquiryType = {
  service_quote: 'service_quote',
  product_order: 'product_order',
  general: 'general',
  emergency: 'emergency',
} as const;

export interface InquiryInput {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  county?: string | null;
  inquiryType: InquiryInputInquiryType;
  /** @nullable */
  serviceId?: number | null;
  message: string;
}

export interface Inquiry {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  county?: string | null;
  inquiryType: string;
  /** @nullable */
  serviceId?: number | null;
  message: string;
  createdAt: string;
}

export interface Project {
  id: number;
  title: string;
  slug: string;
  county: string;
  town: string;
  region: string;
  clientType: string;
  serviceType: string;
  /** @nullable */
  depth?: number | null;
  /** @nullable */
  yield?: string | null;
  /** @nullable */
  duration?: string | null;
  completionYear: number;
  description: string;
  /** @nullable */
  challenge?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
}

export interface ProjectPage {
  items: Project[];
  total: number;
  page: number;
  limit: number;
}

export interface CatalogSummary {
  totalServices: number;
  totalProducts: number;
  productCategories: number;
  serviceCategories: number;
  countiesServed: number;
  yearsExperience?: number;
  projectsCompleted?: number;
}

export type ListServicesParams = {
category?: string;
featured?: boolean;
};

export type ListProductsParams = {
category?: string;
inStock?: boolean;
search?: string;
minPrice?: number;
maxPrice?: number;
page?: number;
limit?: number;
};

export type GetCartParams = {
sessionId: string;
};

export type ListProjectsParams = {
county?: string;
serviceType?: string;
clientType?: string;
featured?: boolean;
page?: number;
limit?: number;
};