# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Borehole Drilling Services Kenya API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: services
    description: Drilling and related services
  - name: products
    description: Borehole equipment and goods
  - name: cart
    description: Shopping cart
  - name: inquiries
    description: Service inquiries and contact
  - name: catalog
    description: Catalog summary endpoints
  - name: projects
    description: Completed project portfolio

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /products/categories:
    get:
      operationId: listProductCategories
      tags: [products]
      summary: List all product categories with counts
      responses:
        "200":
          description: Product categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySummary"

  /products/featured:
    get:
      operationId: listFeaturedProducts
      tags: [products]
      summary: List featured/popular products
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /inquiries:
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Submit a service inquiry or quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryInput"
      responses:
        "201":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List completed projects
      parameters:
        - name: county
          in: query
          schema:
            type: string
        - name: serviceType
          in: query
          schema:
            type: string
        - name: clientType
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Paginated project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPage"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

  /catalog/summary:
    get:
      operationId: getCatalogSummary
      tags: [catalog]
      summary: Get catalog summary stats for homepage
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogSummary"

  /catalog/service-categories:
    get:
      operationId: listServiceCategories
      tags: [catalog]
      summary: List service categories with counts
      responses:
        "200":
          description: Service categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Service:
      type: object
      required: [id, name, slug, category, shortDescription, description, featured]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        priceFrom:
          type: ["number", "null"]
        priceUnit:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        highlights:
          type: array
          items:
            type: string
        coverageAreas:
          type: array
          items:
            type: string

    Product:
      type: object
      required: [id, name, slug, category, description, price, inStock]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        brand:
          type: ["string", "null"]
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        inStock:
          type: boolean
        stockQty:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        specifications:
          type: object
          additionalProperties:
            type: string
        deliveryAvailable:
          type: boolean
        deliveryDays:
          type: ["integer", "null"]
        whatsappOrderEnabled:
          type: boolean

    ProductPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    CategorySummary:
      type: object
      required: [name, count]
      properties:
        name:
          type: string
        count:
          type: integer
        imageUrl:
          type: ["string", "null"]

    Cart:
      type: object
      required: [sessionId, items, total]
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer

    CartItem:
      type: object
      required: [id, productId, productName, price, quantity, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number

    CartItemInput:
      type: object
      required: [sessionId, productId, quantity]
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer

    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    InquiryInput:
      type: object
      required: [name, phone, inquiryType, message]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        inquiryType:
          type: string
          enum: [service_quote, product_order, general, emergency]
        serviceId:
          type: ["integer", "null"]
        message:
          type: string

    Inquiry:
      type: object
      required: [id, name, phone, inquiryType, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        inquiryType:
          type: string
        serviceId:
          type: ["integer", "null"]
        message:
          type: string
        createdAt:
          type: string

    Project:
      type: object
      required: [id, title, slug, county, town, region, clientType, serviceType, completionYear, description, featured]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        county:
          type: string
        town:
          type: string
        region:
          type: string
        clientType:
          type: string
        serviceType:
          type: string
        depth:
          type: ["integer", "null"]
        yield:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        completionYear:
          type: integer
        description:
          type: string
        challenge:
          type: ["string", "null"]
        solution:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]

    ProjectPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    CatalogSummary:
      type: object
      required: [totalServices, totalProducts, productCategories, serviceCategories, countiesServed]
      properties:
        totalServices:
          type: integer
        totalProducts:
          type: integer
        productCategories:
          type: integer
        serviceCategories:
          type: integer
        countiesServed:
          type: integer
        yearsExperience:
          type: integer
        projectsCompleted:
          type: integer