# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borehole Drilling Services Kenya API
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySummary {
  name: string;
  count: number;
  /** @nullable */
  imageUrl?: string | null;
}