# lib/api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borehole Drilling Services Kenya API
 * OpenAPI spec version: 0.1.0
 */
import type { InquiryInputInquiryType } from './inquiryInputInquiryType';

export interface InquiryInput {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  county?: string | null;
  inquiryType: InquiryInputInquiryType;
  /** @nullable */
  serviceId?: number | null;
  message: string;
}