# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borehole Drilling Services Kenya API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpecifications } from './productSpecifications';

export interface Product {
  id: number;
  name: string;
  slug: string;
  category: string;
  /** @nullable */
  brand?: string | null;
  description: string;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  inStock: boolean;
  /** @nullable */
  stockQty?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  specifications?: ProductSpecifications;
  deliveryAvailable?: boolean;
  /** @nullable */
  deliveryDays?: number | null;
  whatsappOrderEnabled?: boolean;
}