# lib/api-zod/src/generated/types/projectPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borehole Drilling Services Kenya API
 * OpenAPI spec version: 0.1.0
 */
import type { Project } from './project';

export interface ProjectPage {
  items: Project[];
  total: number;
  page: number;
  limit: number;
}