# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Borehole Drilling Services Kenya API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  slug: string;
  category: string;
  shortDescription: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  /** @nullable */
  priceFrom?: number | null;
  /** @nullable */
  priceUnit?: string | null;
  /** @nullable */
  duration?: string | null;
  highlights?: string[];
  coverageAreas?: string[];
}